# Library/Swagger/docs/swagger.yaml
basePath: /employee
definitions:
  Dto.BonusList:
    properties:
      reward:
        type: integer
      workingPeriodEnd:
        type: integer
      workingPeriodStart:
        type: integer
    type: object
...
  Dto.LeaveBalancePolicyDTO:
    properties:
      autoCutLeaveWeekly:
        type: boolean
      companyId:
        type: integer
      companyName:
        type: string
      createdBy:
        type: string
      createdTime:
        type: string
      id:
        type: integer
      leaveBalanceAccumulation:
        type: boolean
      leaveBalanceBonusByLenghtOfWork:
        type: boolean
      leaveBalanceBonusList:
        items:
          $ref: '#/definitions/Dto.BonusList'
        type: array
      modifiedBy:
        type: string
      modifiedTime:
        type: string
    type: object
...
  Response.RespErrorStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        example: 400
        type: integer
      title:
        example: bad_request
        type: string
    type: object
  Response.RespResultStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: 'Example: Bearer abcdefghijklmnopqrstuvwxyz12345678901'
  title: Employee Service
  version: "1.0"
paths:
...
  /dapr-state/{stateStore}:
    post:
      consumes:
      - application/json
      parameters:
      - description: stateStore
        in: path
        name: stateStore
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      tags:
      - Sample
  /dapr-state/{stateStore}/{key}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: stateStore
        in: path
        name: stateStore
        required: true
        type: string
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      tags:
      - Sample
    get:
      consumes:
      - application/json
      parameters:
      - description: stateStore
        in: path
        name: stateStore
        required: true
        type: string
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      tags:
      - Sample
...
  /leave-balance/policy:
    get:
      consumes:
      - application/json
      parameters:
      - description: companyName
        in: query
        name: companyName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Select All Data
      tags:
      - Leave Balance Policy
    post:
      consumes:
      - application/json
      description: 'Sample Payload: `{"companyId":2,"companyName":"PT. SSSS","autoCutLeaveWeekly":true,"leaveBalanceAccumulation":true,"leaveBalanceBonusByLenghtOfWork":true,"leaveBalanceBonusList":[{"workingPeriodStart":5,"workingPeriodEnd":15,"reward":5},{"workingPeriodStart":10,"workingPeriodEnd":15,"reward":5}]}`'
      parameters:
      - description: Example Payload
        in: body
        name: '"payload"'
        required: true
        schema:
          $ref: '#/definitions/Dto.LeaveBalancePolicyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Save Data
      tags:
      - Leave Balance Policy
  /leave-balance/policy/{id}:
    put:
      consumes:
      - application/json
      description: 'Sample Payload: `{"autoCutLeaveWeekly":false,"leaveBalanceAccumulation":false,"leaveBalanceBonusByLenghtOfWork":false,"leaveBalanceBonusList":[{"workingPeriodStart":10,"workingPeriodEnd":10,"reward":10},{"workingPeriodStart":11,"workingPeriodEnd":11,"reward":11}]}`'
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Example Payload
        in: body
        name: '"payload"'
        required: true
        schema:
          $ref: '#/definitions/Dto.LeaveBalancePolicyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Update Data
      tags:
      - Leave Balance Policy
...
securityDefinitions:
  Bearer:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"